+++ neoforge/src/main/java/fr/tathan/sky_aesthetics/datagen/SkyPropertiesProvider.java
@@ -72,6 +72,7 @@ public class SkyPropertiesProvider implements DataProvider {
                         Optional.of(SkyColorSettings.createDefaultSettings()),
                         List.of(),
                         Optional.empty(),
+                        Optional.empty(),
                         Optional.empty()
                 )
         );